requirements.txt
appnope==0.1.3
asttokens==2.4.1
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
greenlet==3.0.1
ipykernel==6.27.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
six==1.16.0
SQLAlchemy==2.0.23
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.11